+++ demobaharan/src/main/java/ir/baharn/demobaharan/controller/StudentController.java
@@ -1,9 +1,11 @@
 package ir.baharn.demobaharan.controller;
 
 import ir.baharn.demobaharan.model.Person;
+import ir.baharn.demobaharan.model.Role;
 import ir.baharn.demobaharan.model.Student;
 import ir.baharn.demobaharan.service.PersonService;
 import ir.baharn.demobaharan.service.StudentService;
+import jakarta.servlet.http.HttpSession;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -20,13 +22,21 @@ public class StudentController {
     private PersonService personService;
 
     @GetMapping
-    public String list(Model model) {
+    public String list(Model model , HttpSession session) {
+        Role role = (Role) session.getAttribute("role");
+        if (role !=Role.TEACHER && role !=Role.ADMIN) {
+            return "redirect:/access-denied";
+        }
         model.addAttribute("students", studentService.getAll());
         return "students-list";
     }
 
     @GetMapping("/new")
-    public String createForm(Model model) {
+    public String createForm(Model model , HttpSession session) {
+        Role role = (Role) session.getAttribute("role");
+        if (role !=Role.STUDENT && role !=Role.ADMIN) {
+            return "redirect:/access-denied";
+        }
         model.addAttribute("student", new Student());
         model.addAttribute("persons", personService.getAll());
         return "student-form";
@@ -35,7 +45,12 @@ public class StudentController {
     @PostMapping
     public String save(@RequestParam("personId") Long personId,
                        @RequestParam("studentCode") String studentCode,
-                       @RequestParam("proMajor") String proMajor) {
+                       @RequestParam("proMajor") String proMajor , HttpSession session) {
+
+        Role role = (Role) session.getAttribute("role");
+        if (role !=Role.STUDENT && role !=Role.ADMIN) {
+            return "redirect:/access-denied";
+        }
         Person person = personService.getEntityById(personId);
 
         Student student = new Student();
@@ -48,7 +63,11 @@ public class StudentController {
     }
 
     @GetMapping("/delete/{id}")
-    public String delete(@PathVariable Long id) {
+    public String delete(@PathVariable Long id , HttpSession session) {
+        Role role = (Role) session.getAttribute("role");
+        if (role !=Role.STUDENT && role !=Role.ADMIN) {
+            return "redirect:/access-denied";
+        }
         studentService.delete(id);
         return "redirect:/students";
     }
